Add recomputation of filter frequencies
The real update frequencies of both the package main loop and the imu
loop differ from the configured ones. This difference can be quite big
and have an impact on the response of the filters.

This commit adds a periodic check and recomputation of the filters once
a significant difference in the frequencies is detected.

Fix: Add recalculation of filter constants according to actual measured
 loop frequency
 >
 Due to technical details actual frequencies at which the
 package is running can differ significantly from the configured ones,
 this ensures the filtering is consistent and accurate.