Switch main loop sleep to ticks, add loop time compensation
The VESC_IF->sleep_us() function rounds the interval up to the nearest
tick. The newly calculated tick amount is rounded down, resulting in
sleeping shorter amounts and the actual frequency being closer to the
target. (we then still convert the ticks to us and use
VESC_IF->sleep_us(), as VESC_IF->sleep_ticks() was only added in 6.05)